修复 修改群昵称的 QQ 问题

+++ src/com/windowx/miraibot/PluginMain.java
@@ -646,7 +646,7 @@ public class PluginMain {
 			case "nameCard":
 				if (cmd.length > 2) {
 					try {
-						NormalMember member = group.get(Integer.parseInt(cmd[1]));
+						NormalMember member = group.get(Long.parseLong(cmd[1]));
 						if (member != null) {
 							member.setNameCard(cmd[2]);
 						} else {
